--- README.md
@@ -1,16 +1,0 @@
-This is The Fourth Age Client by Eldarion.
-If you want to compile it the following libraries are required:
-
-- GLFW
-- GLEW
-- SDL2
-
-To compile it type in:
-make
-
-To install it:
-sudo make install
-
-To clean up the workspace:
-make clean
-
