--- engine/graphics/loaders/animation.c
@@ -1,1 +1,0 @@
-#include "animation.h"
