--- engine/graphics/loaders/animation.h
@@ -1,4 +1,0 @@
-#ifndef THE_FOURTH_AGE_ANIMATION_H
-#define THE_FOURTH_AGE_ANIMATION_H
-
-#endif //THE_FOURTH_AGE_ANIMATION_H
