--- game/gameobject.c
@@ -1,1 +1,0 @@
-#include "gameobject.h"
