+++ libs/logger.h
@@ -1,0 +1,11 @@
+#ifndef FOURTH_AGE_CLIENT_LOGGER
+#define FOURTH_AGE_CLIENT_LOGGER
+
+/* For logging. */
+#include <stdio.h>
+
+void    LOG_info    (const char*    text, const char*   infoText);
+
+void    LOG_error   (const char*    text, const char*   errorText);
+
+#endif
