--- libs/maths/quasar/color/colorrgb.h
@@ -1,15 +1,0 @@
-typedef struct {
-	
-	unsigned char r;
-	unsigned char g;
-	unsigned char b;
-	
-} QsrColorRGB;
-
-void qsrInitColorRGB(QsrColorRGB* toInit) {
-	
-	toInit->r = 0;
-	toInit->g = 0;
-	toInit->b = 0;
-}
-
