--- libs/maths/quasar/color/colorrgba.h
@@ -1,17 +1,0 @@
-typedef struct {
-	
-	unsigned char r;
-	unsigned char g;
-	unsigned char b;
-	unsigned char a;
-	
-} QsrColorRGBA;
-
-void qsrInitColorRGBA(QsrColorRGBA* toInit) {
-	
-	toInit->r = 0;
-	toInit->g = 0;
-	toInit->b = 0;
-	toInit->a = 0;
-}
-
