--- libs/maths/quasar/math/todegrees.c
@@ -1,12 +1,0 @@
-#include "todegrees.h"
-
-float qsrRadToDegf(float radians) {
-	
-	return radians * (180 / 3.1415926535);
-}
-
-int qsrRadToDegi(int radians) {
-	
-	return (int) (radians * (180 / 3.1415926535));
-}
-
