--- libs/maths/quasar/math/todegrees.h
@@ -1,5 +1,0 @@
-
-
-float qsrRadToDegf(float degrees);
-int qsrRadToDegi(int degrees);
-
