--- libs/maths/quasar/math/toradians.h
@@ -1,5 +1,0 @@
-
-
-float qsrDegToRadf(float degrees);
-int qsrDegToRadi(int degrees);
-
