--- libs/maths/quasar/vector/vec2.c
@@ -1,20 +1,0 @@
-#include "vec2.h"
-
-void qsrInitVec2f(QsrVec2f* vec) {
-	
-	vec->x = 0.0f;
-	vec->y = 0.0f;
-}
-
-void qsrInitVec2i(QsrVec2i* vec) {
-	
-	vec->x = 0;
-	vec->y = 0;
-}
-
-void qsrInitVec2u(QsrVec2u* vec) {
-	
-	vec->x = 0;
-	vec->y = 0;
-}
-
