--- libs/maths/quasar/vector/vec2.h
@@ -1,33 +1,0 @@
-
-
-typedef struct {
-	
-	float x;
-	float y;
-	
-} QsrVec2f;
-
-void qsrInitVec2f(QsrVec2f* vec);
-
-
-
-typedef struct {
-	
-	int x;
-	int y;
-	
-} QsrVec2i;
-
-void qsrInitVec2i(QsrVec2i* vec);
-
-
-
-typedef struct {
-	
-	unsigned int x;
-	unsigned int y;
-	
-} QsrVec2u;
-
-void qsrInitVec2u(QsrVec2u* vec);
-
