--- libs/maths/quasar/vector/vec4.c
@@ -1,26 +1,0 @@
-#include "vec4.h"
-
-void qsrInitVec4f(QsrVec4f* vec) {
-	
-	vec->x = 0.0f;
-	vec->y = 0.0f;
-	vec->z = 0.0f;
-	vec->w = 0.0f;
-}
-
-void qsrInitVec4i(QsrVec4i* vec) {
-	
-	vec->x = 0;
-	vec->y = 0;
-	vec->z = 0;
-	vec->w = 0;
-}
-
-void qsrInitVec4u(QsrVec4u* vec) {
-	
-	vec->x = 0;
-	vec->y = 0;
-	vec->z = 0;
-	vec->w = 0;
-}
-
