--- libs/maths/quasar/vector/vec4.h
@@ -1,39 +1,0 @@
-
-
-typedef struct {
-	
-	float x;
-	float y;
-	float z;
-	float w;
-	
-} QsrVec4f;
-
-void qsrInitVec4f(QsrVec4f* vec);
-
-
-
-typedef struct {
-	
-	int x;
-	int y;
-	int z;
-	int w;
-	
-} QsrVec4i;
-
-void qsrInitVec4i(QsrVec4i* vec);
-
-
-
-typedef struct {
-	
-	unsigned int x;
-	unsigned int y;
-	unsigned int z;
-	unsigned int w;
-	
-} QsrVec4u;
-
-void qsrInitVec4u(QsrVec4u* vec);
-
